Extract state.py and delete _compute/runtime.py
Move the last two definitions out of braincell/_compute/runtime.py into
braincell/_compute/state.py byte-identically -- NETSTIM_KIND and the
CellRuntimeState dataclass -- and delete runtime.py outright. No
back-compat facade: every consumer now imports from the module that
actually defines the name.

state.py sits at the top of the _compute layer stack, so it imports its
siblings directly (layouts, bindings, bridge) with no TYPE_CHECKING
indirection. Ions are reached only through bindings._build_runtime_nodes.
A TYPE_CHECKING import of Cell resolves the previously dangling "Cell"
forward reference in CellRuntimeState.from_cell.

runtime.py's pass-through __all__ is deleted rather than relocated: six
of its eleven entries named symbols runtime.py did not define. state.py
declares its own __all__ covering only what it defines.

The six-way import in cell.py fans out to each name's true home, merging
clone_morpho into the existing braincell.morph.morphology import and
qualifying cv_value_vector through the already-imported bridge module
(matching the file's 13 other bridge.* call sites) rather than adding a
second import of the same module.

Drops RuntimeModuleAllTest and RuntimeSplitReexportTest (5 tests) -- both
assert the re-export structure this change removes.
