+++ braincell/_compute/runtime_test.py
@@ -2239,7 +2239,7 @@ class IsRootLevelRuntimeNodeUnknownClassTest(unittest.TestCase):
     """Task 18 (C6): unknown channel kinds raise rather than silently return False."""
 
     def test_unknown_channel_kind_raises_value_error(self) -> None:
-        from braincell._compute.runtime import _is_root_level_runtime_node
+        from braincell._compute.bindings import _is_root_level_runtime_node
 
         with self.assertRaises(ValueError) as ctx:
             _is_root_level_runtime_node("channel:__never_registered__")
@@ -2660,7 +2660,7 @@ class PointSynapseRuntimeTest(unittest.TestCase):
 
 from dataclasses import dataclass as _dataclass
 
-from braincell._compute.runtime import (
+from braincell._compute.layouts import (
     CLAMP_KINDS,
     ClampRoutingTable,
     build_clamp_routing_table,
@@ -2982,7 +2982,7 @@ class FnFingerprintWarnsOnOpaqueClosureTest(unittest.TestCase):
     def test_opaque_closure_emits_warning(self) -> None:
         import warnings
 
-        from braincell._compute.runtime import _fn_fingerprint, _opaque_warned
+        from braincell._compute.layouts import _fn_fingerprint, _opaque_warned
 
         class _Opaque:
             __slots__ = ("x",)
@@ -3014,7 +3014,7 @@ class CellRuntimeStateIsMutableTest(unittest.TestCase):
     """ARCH-07: CellRuntimeState is a mutable dataclass; callers must use plain setattr."""
 
     def test_cell_runtime_state_is_not_frozen(self) -> None:
-        from braincell._compute.runtime import CellRuntimeState
+        from braincell._compute.state import CellRuntimeState
 
         self.assertFalse(
             CellRuntimeState.__dataclass_params__.frozen,
@@ -3198,65 +3198,3 @@ class RuntimeSubsolverScheduleTest(unittest.TestCase):
         self.assertEqual(len(nodes), 2)
         self.assertIsNot(nodes[0], nodes[1])
         self.assertEqual({node.substeps for node in nodes}, {2, 3})
-
-
-class RuntimeModuleAllTest(unittest.TestCase):
-    """Pre-split contract: braincell._compute.runtime.__all__ pins public names."""
-
-    def test_all_contains_expected_names(self) -> None:
-        from braincell._compute import runtime as rt
-
-        expected = {
-            "MechanismLayout",
-            "ClampRoutingTable",
-            "build_clamp_routing_table",
-            "CellRuntimeState",
-            "build_placeholder_ions",
-            "clone_morpho",
-            "cv_value_vector",
-            "mechanism_signature",
-        }
-        actual = set(getattr(rt, "__all__", []))
-        missing = expected - actual
-        self.assertFalse(missing, msg=f"missing public symbols: {missing}")
-
-
-class RuntimeSplitReexportTest(unittest.TestCase):
-    """ARCH-02: runtime.py partitions into layouts / state / ions / bindings."""
-
-    def test_mechanism_layout_lives_in_layouts(self) -> None:
-        from braincell._compute import layouts, runtime
-
-        self.assertIs(runtime.MechanismLayout, layouts.MechanismLayout)
-        self.assertIs(runtime.ClampRoutingTable, layouts.ClampRoutingTable)
-        self.assertIs(
-            runtime.build_clamp_routing_table,
-            layouts.build_clamp_routing_table,
-        )
-
-    def test_cell_runtime_state_lives_in_state(self) -> None:
-        from braincell._compute import runtime, state
-
-        self.assertIs(runtime.CellRuntimeState, state.CellRuntimeState)
-
-    def test_ion_helpers_live_in_ions(self) -> None:
-        from braincell._compute import ions, runtime
-
-        self.assertIs(runtime._build_runtime_ions, ions._build_runtime_ions)
-        self.assertIs(runtime._build_default_ions, ions._build_default_ions)
-
-    def test_binding_helpers_live_in_bindings(self) -> None:
-        from braincell._compute import bindings, runtime
-
-        self.assertIs(
-            runtime._resolve_channel_runtime_bindings,
-            bindings._resolve_channel_runtime_bindings,
-        )
-        self.assertIs(
-            runtime._instantiate_runtime_node,
-            bindings._instantiate_runtime_node,
-        )
-        self.assertIs(
-            runtime._BoundIonChannelRuntime,
-            bindings._BoundIonChannelRuntime,
-        )
