+++ braincell/_multi_compartment/probes.py
@@ -25,7 +25,7 @@ import brainstate
 import brainunit as u
 import numpy as np
 
-from braincell._compute.runtime import CellRuntimeState
+from braincell._compute.state import CellRuntimeState
 from braincell.mech import (
     CurrentProbe,
     Density,
